requirements.txt
qumulo_api~=4.1
typing-extensions~=3.10
